kadai.15-1

+++ app/Http/Controllers/TasksController.php
@@ -14,10 +14,19 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        return view('tasks.index',[
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) { 
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        return view('welcome', $data);
+
     }
 
     /**
@@ -46,10 +55,18 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+        
+        /*
         $task = new Task;
+        $task->user_id = \Auth::id();
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
+        */
 
         return redirect('/');
     }
@@ -63,10 +80,14 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = task::findOrFail($id);
-
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id){
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        } else {
+            return redirect('/');
+        }
+        
     }
 
     /**
@@ -78,10 +99,13 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::findOrFail($id);
-        
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id){
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -93,18 +117,24 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
+         $task = Task::findOrFail($id);
+         
+        if (\Auth::id() === $task->user_id){
+            $request->validate([
+                'content' => 'required',
+                'status' => 'required|max:10',
+            ]);
+            
+            $task = Task::findOrFail($id);
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+            
+            return redirect('/');
         
-         $request->validate([
-            'content' => 'required',
-            'status' => 'required|max:10',
-        ]);
-        
-        $task = Task::findOrFail($id);
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-        
-        return redirect('/');
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -116,8 +146,15 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
-        $task->delete();
-        return redirect('/');
+        
+        if (\Auth::id() === $task->user_id){
+            $task->delete();
+            
+            return redirect('/');
+            
+        } else {
+            return redirect('/');
+        }
     }
 }
